requirements.txt
annotated-types==0.7.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
deprecation==2.1.0
flatbuffers==25.12.19
idna==3.11
lance-namespace==0.5.2
lance-namespace-urllib3-client==0.5.2
lancedb==0.29.2
markdown-it-py==4.0.0
mdurl==0.1.2
mpmath==1.3.0
muna==0.0.91
numpy==2.4.2
onnxruntime==1.24.2
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
protobuf==6.33.5
pyarrow==23.0.1
pyclipper==1.4.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.27.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
